+++ backend/routers/children.py
@@ -125,10 +125,10 @@ def _annotate_history(q_answers, points_per_clear, stage: int = 1):
     return history
 
 
-def _get_points_per_clear(db: Session) -> int:
+def _get_points_per_clear(db: Session) -> float:
     s = db.query(Setting).get("points_per_clear")
     try:
-        return int(s.value) if s else 2
+        return float(s.value) if s else 2
     except Exception:
         return 2
